move default error handling into flow package
